Rat's Unit Pack
The first LOUD Extra is the start of my personal compilation of units. This will include equipment that I think may be interesting to use but am not giving consideration for inclusion into the LOUD Standard. If, in the future, I make units for the Standard, they will go into LOUD Unit Additions. Otherwise, they'll go here.

For now, it's just going to be inclusions from the outside. At the moment, there's only the Ennui; a Cybran T3 "Prototype Multi Missile Submarine". This comes from the Antares Unit Pack by AntaresOne. Its ability to fire with practically perfect accuracy from underwater may be overpowered, but according to the gossip on the team, Cybran is already IMBA anyway.

Be aware that AntaresOne's permission was not granted for this redistribution; they have not been online on ModDB all year. I sent a message asking permission as a token of good faith and for posterity. If it becomes an issue, I'll deal with it.