RatUnits: AC-500 Terror High Support Gunship
From Antares Unit Pack 3. For all intents and purposes, it's an AC-130 "Spooky" for the UEF at T4. Strictly speaking, it is a "gunship", but is winged unlike any other unit normally designated as such. Seeing as how nothing exists like this in the game, it seems to me that a unit like this would nicely differentiate the UEF's experimental air offerings from the CZAR (which has similarities to the Citadels), the Ahwassa (which has similarities to the Lucidity), and the Soul Rippers.