RatUnits: Imperator (UEF Experimental Artillery)
On request by deathtrooper900#3119. The Imperator is a T4 static artillery installation not unlike the Mavor. It differs in firing a 20-round salvo over a smaller range of 52km (vs. Mavor's 80).

The only modifications were to tidy up its frontend strings and change its MuzzleVelocity and MuzzleVelocityReduceDistance from 100 and 2550 to 120 and 2200, respectively. Their original values prevented it from even attempting to engage targets in certain parts of its range.

Also note that because this overflows the previous custom unit file, it was split it into two.